requirements.txt
asn1crypto==0.24.0
buttercms-python==1.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-bootstrap3==10.0.1
gunicorn==19.9.0
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
six==1.11.0
static3==0.7.0
urllib3==1.23
